tools/normalizer/requirements.txt
jsonlines==4.0.0
pyyaml==6.0.2
pandas==2.2.3